+++ presidio-analyzer-new/analyzer/predefined_recognizers/__init__.py
@@ -1,0 +1,1 @@
+from .credit_card_recognizer import CreditCardRecognizer
