+++ presidio-analyzer-new/tests/test_pattern_recognizer.py
@@ -3,8 +3,7 @@ import os
 
 # https://www.datatrans.ch/showcase/test-cc-numbers
 # https://www.freeformatter.com/credit-card-number-generator-validator.html
-from analyzer.pattern import Pattern
-from analyzer.pattern_recognizer import PatternRecognizer
+from analyzer import PatternRecognizer, Pattern
 
 
 class TestRecognizer(PatternRecognizer):
